improved setup and data_cleaning modules
1_setup.R now includes a config file driven option to load a custom vulnerability and outcome generating script which should be called directly in the 2_data_cleaning.R module.  Improved error handling for importing raw survey files.  Updated template for the config - template.yml file.  Misc helper scripts added to the adhoc_scripts folder

+++ analyses/a_new_analysis_template/1_import_data.R
@@ -9,22 +9,43 @@
 ###################################
 # READ IN SURVEY | DHS | PATHWAYS
 
-IR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_ir_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
-BR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_br_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
-KR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_kr_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
-HH <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_hh_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
-MR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_mr_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
+# WOMAN'S RECODE
+if (file.exists(paste0(data_path, config::get("dhs_ir_file")))) {
+  IR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_ir_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
+}
+
+# BIRTH RECODE
+if (file.exists(paste0(data_path, config::get("dhs_br_file")))) {
+  BR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_br_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
+}
+
+# CHILD RECODE
+if (file.exists(paste0(data_path, config::get("dhs_kr_file")))) {
+  KR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_kr_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
+}
+
+# HOUSEHOLD RECODE
+if (file.exists(paste0(data_path, config::get("dhs_hh_file")))) {
+  HH <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_hh_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
+}
+
+# MEN'S RECODE
+if (file.exists(paste0(data_path, config::get("dhs_mr_file")))) {
+  MR <- data.table(read.dta13(file = paste0(data_path, config::get("dhs_mr_file")), fromEncoding="utf-8")) %>% dplyr::mutate(survey = config::get("survey_name"))
+}
+
+
 # survey <-
 
 
 ###################################
 # SAVE AS RDS
 
-saveRDS(IR, file = paste0(data_path, "IR.rds"))
-saveRDS(BR, file = paste0(data_path, "BR.rds"))
-saveRDS(KR, file = paste0(data_path, "KR.rds"))
-saveRDS(HH, file = paste0(data_path, "HH.rds"))
-saveRDS(MR, file = paste0(data_path, "MR.rds"))
+if (exists("IR")) {saveRDS(IR, file = paste0(data_path, "IR.rds"))}
+if (exists("BR")) {saveRDS(BR, file = paste0(data_path, "BR.rds"))}
+if (exists("KR")) {saveRDS(KR, file = paste0(data_path, "KR.rds"))}
+if (exists("HH")) {saveRDS(HH, file = paste0(data_path, "HH.rds"))}
+if (exists("MR")) {saveRDS(MR, file = paste0(data_path, "MR.rds"))}
 # saveRDS(survey, file = )
